fix: scoping shortcuts for triage modals
