+++ packages/tegon-sdk/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tegonhq/sdk",
-  "version": "0.1.11",
+  "version": "0.1.12",
   "description": "Tegon Node.JS SDK",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
